fix: slider.dart 분리
